Upload function for multi files implemented

+++ src/main/java/com/spring/project/board/controller/BoardControllerImpl.java
@@ -1,6 +1,7 @@
 package com.spring.project.board.controller;
 
 import java.io.File;
+import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -28,6 +29,7 @@ import org.springframework.web.servlet.ModelAndView;
 
 import com.spring.project.board.service.BoardService;
 import com.spring.project.board.vo.ArticleVO;
+import com.spring.project.board.vo.ImageVO;
 import com.spring.project.member.vo.MemberVO;
 
 @Controller
@@ -58,8 +60,10 @@ public class BoardControllerImpl implements BoardController {
 	@ResponseBody
 	public ResponseEntity addNewArticle(MultipartHttpServletRequest multipartRequest, HttpServletResponse response)
 			throws Exception {
-		logger.info("addNewArticle() called");			
+		logger.info("addNewArticle() called");	
 		multipartRequest.setCharacterEncoding("utf-8");
+//		String imageFileName = null;
+		
 		Map<String, Object> articleMap = new HashMap<String, Object>();
 		Enumeration enu = multipartRequest.getParameterNames();
 		while(enu.hasMoreElements()) {
@@ -68,38 +72,56 @@ public class BoardControllerImpl implements BoardController {
 			articleMap.put(name, value);
 		}
 		
-		String imageFileName = upload(multipartRequest);
+//		String imageFileName = upload(multipartRequest);
 		HttpSession session = multipartRequest.getSession();
 		MemberVO memberVO = (MemberVO) session.getAttribute("member");
 		String id = memberVO.getId();
-		articleMap.put("parentNO", 0);
+//		articleMap.put("parentNO", 0);
 		articleMap.put("id", id);
-		articleMap.put("imageFileName", imageFileName);
+//		articleMap.put("imageFileName", imageFileName);
+		
+		List<String> fileList = upload(multipartRequest);
+		List<ImageVO> imageFileList = new ArrayList<ImageVO>();
+		if(fileList != null && fileList.size() != 0) {
+			for(String fileName : fileList) {
+				ImageVO imageVO = new ImageVO();
+				imageVO.setImageFileName(fileName);
+				imageFileList.add(imageVO);
+			}
+			articleMap.put("imageFileList", imageFileList);
+		}
 		
 		String message;
 		ResponseEntity resEnt;
 		HttpHeaders responseHeaders = new HttpHeaders();
 		responseHeaders.add("Content-type", "text/html; charset=utf-8");
+		String imageFileName;
 		try {
 			int articleNO = boardService.addNewArticle(articleMap);
-			
-			if(imageFileName != null && imageFileName.length() != 0) {
-				// srcFile : upload() 통해 생성된 파일
-				File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
-				File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + articleNO);
-				FileUtils.moveFileToDirectory(srcFile, destDir, true);
-			} 
-			
+			if(imageFileList != null && imageFileList.size() != 0) {
+				// upload() 로직으로 DB 처리된 imageFileList
+				for(ImageVO imageVO : imageFileList) {
+					imageFileName = imageVO.getImageFileName();
+					// srcFile : upload() 통해 생성된 파일								
+					File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
+					File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + articleNO);
+					FileUtils.moveFileToDirectory(srcFile, destDir, true);					
+				}
+			}
 			message = "<script>"
 					+ " alert('글이 정상적으로 저장됐습니다'); "
 					+ " location.href='" + multipartRequest.getContextPath() + "/board/listArticles.do';"
 					+ "</script>";
-			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
-					
+			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);		
 		} catch(Exception e) {
-			// upload() 통해 생성된 srcFile 삭제
-			File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
-			srcFile.delete();
+			if(imageFileList != null && imageFileList.size() != 0) {
+				for(ImageVO imageVO : imageFileList) {
+					imageFileName = imageVO.getImageFileName();
+					// upload() 통해 생성된 srcFile 삭제			
+					File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
+					srcFile.delete();					
+				}
+			}
 			
 			message = "<script>"
 					+ " alert('글 작성에 실패했습니다!'); "
@@ -108,9 +130,92 @@ public class BoardControllerImpl implements BoardController {
 			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
 			e.printStackTrace();
 		}
-		
 		return resEnt;
 	}
+
+////////////////////////////// 첨부파일 한 개일 때 사용한 addNewArticle() ///////////////////////////////////
+//	@Override
+//	@RequestMapping(value="/board/addNewArticle.do", method=RequestMethod.POST)
+//	@ResponseBody
+//	public ResponseEntity addNewArticle(MultipartHttpServletRequest multipartRequest, HttpServletResponse response)
+//			throws Exception {
+//		logger.info("addNewArticle() called");			
+//		multipartRequest.setCharacterEncoding("utf-8");
+//		Map<String, Object> articleMap = new HashMap<String, Object>();
+//		Enumeration enu = multipartRequest.getParameterNames();
+//		while(enu.hasMoreElements()) {
+//			String name = (String) enu.nextElement();
+//			String value = multipartRequest.getParameter(name);
+//			articleMap.put(name, value);
+//		}
+//		
+//		String imageFileName = upload(multipartRequest);
+//		HttpSession session = multipartRequest.getSession();
+//		MemberVO memberVO = (MemberVO) session.getAttribute("member");
+//		String id = memberVO.getId();
+//		articleMap.put("parentNO", 0);
+//		articleMap.put("id", id);
+//		articleMap.put("imageFileName", imageFileName);
+//		
+//		String message;
+//		ResponseEntity resEnt;
+//		HttpHeaders responseHeaders = new HttpHeaders();
+//		responseHeaders.add("Content-type", "text/html; charset=utf-8");
+//		try {
+//			int articleNO = boardService.addNewArticle(articleMap);
+//			
+//			if(imageFileName != null && imageFileName.length() != 0) {
+//				// srcFile : upload() 통해 생성된 파일
+//				File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
+//				File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + articleNO);
+//				FileUtils.moveFileToDirectory(srcFile, destDir, true);
+//			} 
+//			
+//			message = "<script>"
+//					+ " alert('글이 정상적으로 저장됐습니다'); "
+//					+ " location.href='" + multipartRequest.getContextPath() + "/board/listArticles.do';"
+//					+ "</script>";
+//			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+//					
+//		} catch(Exception e) {
+//			// upload() 통해 생성된 srcFile 삭제
+//			File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
+//			srcFile.delete();
+//			
+//			message = "<script>"
+//					+ " alert('글 작성에 실패했습니다!'); "
+//					+ " location.href='" + multipartRequest.getContextPath() + "/board/articleForm.do';"
+//					+ "</script>";
+//			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+//			e.printStackTrace();
+//		}
+//		
+//		return resEnt;
+//	}	
+	
+//////////////////////////////첨부파일 한 개일 때 사용한 upload() ///////////////////////////////////
+//	private String upload(MultipartHttpServletRequest multipartRequest) throws Exception {
+//		logger.info("upload() called");		
+//		String imageFileName = null;
+//		Iterator<String> fileNames = multipartRequest.getFileNames();
+//		
+//		while(fileNames.hasNext()) {
+//			String fileName = fileNames.next();
+//			MultipartFile mFile = multipartRequest.getFile(fileName);
+//			imageFileName = mFile.getOriginalFilename();
+//			File file = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + fileName);
+//			if(mFile.getSize() != 0) {
+//				file.getParentFile().mkdir();
+//				mFile.transferTo(new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName));
+//			}
+//			
+//			logger.info("upload() called");
+//			logger.info("fileName : " + fileName);
+//			logger.info("imageFileName : " + imageFileName);
+//		}
+//	return imageFileName;
+//	}
+//	
 	
 	@RequestMapping(value="/board/*Form.do", method=RequestMethod.GET)
 	public ModelAndView form(HttpServletRequest request, HttpServletResponse response) {
@@ -121,26 +226,28 @@ public class BoardControllerImpl implements BoardController {
 		return mav;
 	}
 	
-	private String upload(MultipartHttpServletRequest multipartRequest) throws Exception {
+	private List<String> upload(MultipartHttpServletRequest multipartRequest) throws Exception {
 		logger.info("upload() called");		
-		String imageFileName = null;
+		List<String> fileList = new ArrayList<String>();
 		Iterator<String> fileNames = multipartRequest.getFileNames();
 		
 		while(fileNames.hasNext()) {
 			String fileName = fileNames.next();
 			MultipartFile mFile = multipartRequest.getFile(fileName);
-			imageFileName = mFile.getOriginalFilename();
+			String originalFileName = mFile.getOriginalFilename();
+			fileList.add(originalFileName);
 			File file = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + fileName);
 			if(mFile.getSize() != 0) {
-				file.getParentFile().mkdir();
-				mFile.transferTo(new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName));
+				if(!file.exists()) {
+					file.getParentFile().mkdir();
+					mFile.transferTo(new File(ARTICLE_IMAGE_REPO + "\\temp\\" + originalFileName));
+				}
 			}
-			
 			logger.info("upload() called");
 			logger.info("fileName : " + fileName);
-			logger.info("imageFileName : " + imageFileName);
+			logger.info("originalFileName : " + originalFileName);
 		}
-	return imageFileName;
+	return fileList;
 	}
 
 	@Override
@@ -170,13 +277,16 @@ public class BoardControllerImpl implements BoardController {
 			String value = multipartRequest.getParameter(name);
 			articleMap.put(name, value);
 		}
-		
-		String imageFileName = upload(multipartRequest);
-		HttpSession session = multipartRequest.getSession();
-		MemberVO memberVO = (MemberVO) session.getAttribute("member");
-		String id = memberVO.getId();
-		articleMap.put("id", id);
-		articleMap.put("imageFileName", imageFileName);
+		List<String> fileList = upload(multipartRequest);
+		List<ImageVO> imageFileList = new ArrayList<ImageVO>();
+		if(fileList != null && fileList.size() != 0) {
+			for(String fileName : fileList) {
+				ImageVO imageVO = new ImageVO();
+				imageVO.setImageFileName(fileName);
+				imageFileList.add(imageVO);
+			}
+			articleMap.put("imageFileList", imageFileList);
+		}		
 				
 		String articleNO = (String) articleMap.get("articleNO");
 		String message;
@@ -184,13 +294,17 @@ public class BoardControllerImpl implements BoardController {
 		HttpHeaders responseHeaders = new HttpHeaders();
 		responseHeaders.add("Content-Type", "text/html; charset=utf-8");
 		
+		String imageFileName;
 		try {
 			boardService.modArticle(articleMap);
-			if(imageFileName != null && imageFileName.length() != 0) {
-				File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
-				File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + articleNO);
-				FileUtils.moveFileToDirectory(srcFile, destDir, true);
-				
+			if(imageFileList != null && imageFileList.size() != 0) {
+				for(ImageVO imageVO : imageFileList) {
+					imageFileName = imageVO.getImageFileName();
+					// srcFile : upload() 통해 생성된 파일								
+					File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
+					File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + articleNO);
+					FileUtils.moveFileToDirectory(srcFile, destDir, true);					
+				}			
 				String originalFileName = (String) articleMap.get("originalFileName");
 				File oldFile = new File(ARTICLE_IMAGE_REPO + "\\" + articleNO + "\\" + originalFileName);
 				oldFile.delete();
@@ -201,14 +315,110 @@ public class BoardControllerImpl implements BoardController {
 					+ " </script> ";
 			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
 		} catch(Exception e) {
-			File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
-			srcFile.delete();
+			if(imageFileList != null && imageFileList.size() != 0) {
+				for(ImageVO imageVO : imageFileList) {
+					imageFileName = imageVO.getImageFileName();
+					File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
+					srcFile.delete();
+				}
+			}
 			message = "<script>"
 					+ " alert('글 수정 과정에서 에러가 발생했습니다!'); "
 					+ " location.href='" + multipartRequest.getContextPath() + "/board/viewArticle?ArticleNO=" + articleNO + "';"
 					+ " </script> ";
 			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
 				e.printStackTrace();
+		}
+		return resEnt;
+	}	
+	
+//////////////////////////////첨부파일 한 개일 때 사용한 modArticle() ///////////////////////////////////
+//	@Override
+//	@RequestMapping(value="/board/modArticle.do", method=RequestMethod.POST)
+//	@ResponseBody
+//	public ResponseEntity modArticle(MultipartHttpServletRequest multipartRequest, HttpServletResponse response)
+//			throws Exception {
+//		logger.info("modArticle() called");
+//		multipartRequest.setCharacterEncoding("utf-8");
+//		Map<String, Object> articleMap = new HashMap<String, Object>();
+//		Enumeration enu = multipartRequest.getParameterNames();
+//		while(enu.hasMoreElements()) {
+//			String name = (String) enu.nextElement();
+//			String value = multipartRequest.getParameter(name);
+//			articleMap.put(name, value);
+//		}
+//		
+//		String imageFileName = upload(multipartRequest);
+//		HttpSession session = multipartRequest.getSession();
+//		MemberVO memberVO = (MemberVO) session.getAttribute("member");
+//		String id = memberVO.getId();
+//		articleMap.put("id", id);
+//		articleMap.put("imageFileName", imageFileName);
+//				
+//		String articleNO = (String) articleMap.get("articleNO");
+//		String message;
+//		ResponseEntity resEnt = null;
+//		HttpHeaders responseHeaders = new HttpHeaders();
+//		responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+//		
+//		try {
+//			boardService.modArticle(articleMap);
+//			if(imageFileName != null && imageFileName.length() != 0) {
+//				File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
+//				File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + articleNO);
+//				FileUtils.moveFileToDirectory(srcFile, destDir, true);
+//				
+//				String originalFileName = (String) articleMap.get("originalFileName");
+//				File oldFile = new File(ARTICLE_IMAGE_REPO + "\\" + articleNO + "\\" + originalFileName);
+//				oldFile.delete();
+//			} 
+//			message = "<script>"
+//					+ " alert('글이 정상적으로 수정됐습니다'); "
+//					+ " location.href='" + multipartRequest.getContextPath() + "/board/viewArticle?ArticleNO=" + articleNO + "';"
+//					+ " </script> ";
+//			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+//		} catch(Exception e) {
+//			File srcFile = new File(ARTICLE_IMAGE_REPO + "\\temp\\" + imageFileName);
+//			srcFile.delete();
+//			message = "<script>"
+//					+ " alert('글 수정 과정에서 에러가 발생했습니다!'); "
+//					+ " location.href='" + multipartRequest.getContextPath() + "/board/viewArticle?ArticleNO=" + articleNO + "';"
+//					+ " </script> ";
+//			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+//				e.printStackTrace();
+//		}
+//		return resEnt;
+//	}
+
+	@Override
+	@RequestMapping(value="board/removeArticle.do", method=RequestMethod.POST)
+	@ResponseBody
+	public ResponseEntity removeArticle(int articleNO, 
+			HttpServletRequest request, HttpServletResponse response) throws Exception {
+		logger.info("removeArticle() called");
+		response.setContentType("text/html; charset=utf-8");
+		ResponseEntity resEnt = null;
+		HttpHeaders responseHeaders = new HttpHeaders();
+		responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+		String message = "";
+		
+		try {
+			boardService.removeArticle(articleNO);
+			File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + articleNO);
+			FileUtils.deleteDirectory(destDir);
+			
+			message = "<script>"
+					+ " alert('글을 삭제했습니다!'); "
+					+ " location.href='" + request.getContextPath() + "/board/listArticles.do';"
+					+ " </script> ";
+			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);			
+		} catch(Exception e) {
+			message = "<script>"
+					+ " alert('오류 발생! 글 삭제 실패'); "
+					+ " location.href='" + request.getContextPath() + "/board/listArticles.do';"
+					+ " </script> ";
+			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+			e.printStackTrace();
 		}
 		return resEnt;
 	}
